services/recommend/requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
grpcio==1.66.1
protobuf==5.27.3
httpx==0.27.2
